pyproject.toml
[project]
name = "draft_mtg"
version = "1.0.0"
description = ""
authors = [{ name = "Filip Karpinski", email = "F.Karpinski@o2.pl" }]
requires-python = "~=3.12"
readme = "README.md"
dependencies = [
    "pydantic>=2.10.3,<3",
    "typer>=0.15.1,<0.16",
    "fastapi>=0.115.11,<0.116",
    "sqlalchemy[asyncio]>=2.0.40,<3",
    "uvicorn>=0.34.0,<0.35",
    "pre-commit>=4.1.0,<5",
    "alembic>=1.15.1,<2",
    "psycopg2-binary>=2.9.10,<3",
    "python-multipart>=0.0.20,<0.0.21",
    "passlib>=1.7.4,<2",
    "pyjwt>=2.10.1,<3",
    "python-dotenv>=1.0.0,<2",
    "ruff>=0.11.2,<0.12",
    "asyncpg>=0.30.0,<0.31",
    "pytest>=8.3.5,<9",
    "pytest-asyncio>=0.26.0,<0.27",
    "pydantic-settings>=2.9.1,<3",
    "email-validator>=2.2.0,<3",
]

[dependency-groups]
dev = [
    "pylint==3.2.1",
    "pytest-cov==5",
    "coverage==7.5.1",
    "pytest>=8.3.4,<9",
    "pytest-mock>=3.14.0,<4",
    "mypy>=1.13.0,<2",
    "ipykernel>=6.29.5,<7",
]

[tool.hatch.build.targets.sdist]
include = ["app"]

[tool.hatch.build.targets.wheel]
include = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_untyped_calls = true

[[tool.mypy.overrides]]
module = "app.alembic.*"
ignore_errors = true
follow_imports = "skip"

[tool.autoflake]
remove_all_unused_imports = true
remove_unused_variables = true

[tool.ruff]
line-length = 120
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "app/alembic/*",
]

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["E501", "B008"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

